% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_converter_functions.R
\name{pace_km_to_mi_num_input}
\alias{pace_km_to_mi_num_input}
\title{Pace km to pace mi, number input}
\usage{
pace_km_to_mi_num_input(pace_km_min, pace_km_sec)
}
\arguments{
\item{pace_km_min}{A number with the whole minutes in the pace in minute per kilometer}

\item{pace_km_sec}{A number with the seconds in the pace in minute per kilometer}
}
\value{
A string with the pace in minutes per mile, where the minutes are separated from the seconds by a colon.
}
\description{
This function converts a pace from minutes per kilometer to minutes per mile, taking two numbers as inputs (minutes and seconds) and returning a string as output. Only works with one observation at a time for now.
}
\examples{
pace_km_to_mi_num_input(7,30)
pace_km_to_mi_num_input(5,27)

}
