% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_converter_functions.R
\name{pace_km_to_mi_text}
\alias{pace_km_to_mi_text}
\title{Pace km to pace mi, text}
\usage{
pace_km_to_mi_text(pace_km_text)
}
\arguments{
\item{pace_km_text}{A string or vector of strings with pace in minutes per kilometer, where the minutes are separated from the seconds by a colon.}
}
\value{
A string or vector of strings with the pace in minutes per mile, where the minutes are separated from the seconds by a colon.
}
\description{
This function converts paces from minutes per kilometer to minutes per mile, taking strings as inputs and returning strings as outputs
}
\examples{
pace_km_to_pace_mi_text("6:03")
pace_km_to_pace_mi_text(c("7:30", "5:27", "6:42"))

}
