% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_converter_functions.R
\name{pace_mi_to_km_num_input}
\alias{pace_mi_to_km_num_input}
\title{Pace mi to pace km, number input}
\usage{
pace_mi_to_km_num_input(pace_mi_min, pace_mi_sec)
}
\arguments{
\item{pace_mi_min}{A number with the whole minutes in the pace in minute per mi}

\item{pace_mi_sec}{A number with the seconds in the pace in minute per mi}
}
\value{
A string with the pace in minutes per kilometer, where the minutes are separated from the seconds by a colon.
}
\description{
This function converts a pace from minutes per mile to minutes per kilometer, taking two numbers as inputs (minutes and seconds) and returning a string as output. Only works with one observation at a time for now.
}
\examples{
pace_mi_to_km_num_input(7,30)
pace_mi_to_km_num_input(5,27)

}
