% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_converter_functions.R
\name{pace_mi_to_km_text}
\alias{pace_mi_to_km_text}
\title{Pace mi to pace km, text}
\usage{
pace_mi_to_km_text(pace_mi_text)
}
\arguments{
\item{pace_mi_text}{A string or vector of strings with pace in minutes per mile, where the minutes are separated from the seconds by a colon.}
}
\value{
A string or vector of strings with the pace in minutes per kilometer, where the minutes are separated from the seconds by a colon.
}
\description{
This function converts paces from minutes per mile to minutes per kilometer, taking strings as inputs and returning strings as outputs
}
\examples{
pace_mi_to_km_text("7:45")
pace_mi_to_km_text(c("11:02", "9:00", "10:34"))

}
