% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_converter_functions.R
\name{pace_mi_to_pace_km}
\alias{pace_mi_to_pace_km}
\title{Pace mi to pace km}
\usage{
pace_mi_to_pace_km(pace_mi)
}
\arguments{
\item{pace_mi}{A string or vector of strings with pace in minutes per mile, where the minutes are separated from the seconds by a colon.}
}
\value{
A string or vector of strings with the pace in minutes per kilometer, where the minutes are separated from the seconds by a colon.
}
\description{
This function converts paces from minutes per mile to minutes per kilometer
}
\examples{
pace_mi_to_pace_km("7:45")
pace_mi_to_pace_km(c("11:02", "9:00", "10:34"))

}
